# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "docs/API_CONTRACT.md \u00a70. The body of every error response, no\nexceptions \u2014 including FastAPI's own framework-generated ones (404 on an\nunrouted path, 422 on a request body pydantic cannot parse), which the\nhandlers below also render into this same shape rather than letting\nStarlette's default plain-text/JSON-detail bodies escape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProblemListOut": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/ProblemListItem"
            },
            "title": "Problems",
            "type": "array"
          }
        },
        "required": [
          "problems"
        ],
        "title": "ProblemListOut",
        "type": "object"
      },
...
      "ProblemStatus": {
        "enum": [
          "open",
          "resolved"
        ],
        "title": "ProblemStatus",
        "type": "string"
      },
      "ProblemType": {
        "enum": [
          "disease",
          "pest"
        ],
        "title": "ProblemType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "SIH26131 \u2014 early detection and management of crop diseases and pest infestations. Four crops, 26 crop-namespaced targets. Wire format is frozen in docs/API_CONTRACT.md.",
    "title": "Bhoomi v3",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/farms/{farm_id}/problems": {
      "get": {
        "description": "Filtered, paginated problem list. Newest first.",
        "operationId": "list_problems_api_v1_farms__farm_id__problems_get",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProblemType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No, or an invalid, bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Either the caller is not a farmer, or that farm belongs to a different farmer's account."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "That farm does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A path or query parameter did not parse."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Problems",
        "tags": [
          "case file"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/timeline": {
      "get": {
        "description": "Chronological case file, newest first.\n\nAssembled in Python from six tables rather than as a UNION: the row shapes\nhave almost nothing in common, and a UNION would need every column widened\nto text. The volume is one farm's history, which is small by construction.",
        "operationId": "farm_timeline_api_v1_farms__farm_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No, or an invalid, bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Either the caller is not a farmer, or that farm belongs to a different farmer's account."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "That farm does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A path or query parameter did not parse."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Farm Timeline",
        "tags": [
          "case file"
        ]
      }
    },
...
    "/api/v1/problems/{problem_id}": {
      "get": {
        "description": "Detail with photos, observations and advisory.\n\nReturns whatever exists and omits what does not. No section is emitted\nempty-but-present, and nothing is invented to fill a shape.",
        "operationId": "get_problem_api_v1_problems__problem_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "problem_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Problem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No, or an invalid, bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not a farmer -- this case file is read by the farmer who owns it (see the module docstring for where agronomist/official read instead)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "That problem does not exist, or belongs to a different account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A path or query parameter did not parse."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Problem",
        "tags": [
          "case file"
        ]
      }
    },
    "/api/v1/problems/{problem_id}/escalate": {
      "post": {
        "description": "Manually escalate this problem to an agronomist. docs/API_CONTRACT.md \u00a712.\n\nA route-surface addition only, no new decision logic: this is the same\nescalate() that already fires automatically from the gate (below-floor,\nout-of-scope, ambiguous-with-no-cue), from a \"got worse\" follow-up\nresponse, and from an inspection-tier alert. Idempotent per problem\n(escalate()'s own guarantee, core/services/escalation.py): calling this\ntwice on the same problem returns the SAME case, not a second one in the\nqueue.",
        "operationId": "escalate_problem_api_v1_problems__problem_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "problem_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Problem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No, or an invalid, bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The caller is not a farmer -- this case file is read by the farmer who owns it (see the module docstring for where agronomist/official read instead)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "That problem does not exist, or belongs to a different account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A path or query parameter did not parse."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Escalate Problem",
        "tags": [
          "case file"
        ]
      }
    },
...
  }
}